@@ -34,9 +34,9 @@
     "@babel/register": "^7.14.5",
     "@tsconfig/recommended": "^1.0.1",
     "@types/jest": "^26.0.24",
-    "@types/react": "^17.0.15",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-test-renderer": "^17.0.1",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "@types/react-test-renderer": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^4.28.5",
     "@typescript-eslint/parser": "^4.28.5",
     "babel-core": "^7.0.0-bridge.0",
@@ -49,23 +49,22 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "jest": "^27.0.6",
     "node-notifier": "^10.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-test-renderer": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-test-renderer": "^18.2.0",
     "ts-loader": "^9.2.4",
     "typescript": "^4.3.5",
-    "webpack": "^5.47.1",
-    "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.11.2"
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.11.1"
   },
   "scripts": {
     "prepublish": "tsc",
     "postversion": "git push && git push --tags",
     "test": "jest",
-    "watch": "webpack serve --inline --hot --progress --open --mode development",
+    "watch": "webpack serve --hot --progress --open --mode development",
     "coverage": "NODE_ENV=test jest --coverage && cat ./coverage/lcov.info | coveralls"
   },
-  "dependencies": {},
   "publishConfig": {
     "access": "public"
   }
